requirements.txt
requests
cryptography
reportlab
pyttsx3
edge-tts
bcrypt
python-dotenv
paramiko
argon2-cffi
flask
flask-cors
flask-limiter
gunicorn
psycopg2-binary
pytest
